Separate auto intialise
