+++ src/behaviors/index.js
@@ -2,3 +2,4 @@ export { default as draggable } from './draggable';
 export { default as droppable } from './droppable';
 export { default as selectable } from './selectable';
 export { default as scrollable } from './scrollable';
+export { default as autoInitialisedForm } from './autoInitialisedForm';
