+++ src/containers/Elements/Form.js
@@ -1,38 +1,10 @@
-import React from 'react';
-import PropTypes from 'prop-types';
 import { reduxForm } from 'redux-form';
-import { fromPairs, map } from 'lodash';
+import { autoInitialisedForm } from '../../behaviors';
 import { Form } from '../../components/Elements';
-
-const typeInitalValue = (field) => {
-  switch (field.type) {
-    case 'checkbox_group':
-      return fromPairs(map(field.options, option => [option, false]));
-    case 'checkbox_list':
-      return [];
-    default:
-      return '';
-  }
-};
-
-const initialValues = fields =>
-  fromPairs(map(fields, field => [field.name, typeInitalValue(field)]));
 
 /**
   * Renders a redux form that contains fields according to a `fields` config.
   */
-const autoInitialisedForm = (WrappedComponent) => {
-  const AutoInitialisedForm = props => (
-    <WrappedComponent initialValues={initialValues(props.fields)} {...props} />
-  );
-
-  AutoInitialisedForm.propTypes = {
-    fields: PropTypes.array.isRequired,
-  };
-
-  return AutoInitialisedForm;
-};
-
 export default autoInitialisedForm(reduxForm({
   destroyOnUnmount: true,
   forceUnregisterOnUnmount: true,
